// src/api/models/UserRead.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UserRead = {
    username: string;
    first_name: string;
    last_name: string;
    email: string;
    id: number;
    created_date: string;
    updated_date?: (string | null);
};